dev(docker): suggest not to use dev containers on macos

+++ docker-compose.local.run.sh
@@ -6,11 +6,20 @@ DANGER_DROP_VOLUMES="false"
 for i in "$@"
 do
 case $i in
+    --stop)
+      STOP_CONTAINERS="true"
+    ;;
     --remove-domifa-containers)
-      DANGER_REMOVE_CONTAINERS="true"
+      STOP_CONTAINERS="true"
+      REMOVE_CONTAINERS="true"
     ;;
     --drop-domifa-volumes)
+      STOP_CONTAINERS="true"
+      REMOVE_CONTAINERS="true"
       DANGER_DROP_VOLUMES="true"
+    ;;
+    --with-dev-containers)
+      WITH_DEV_CONTAINERS="true"
     ;;
     *)
     # unknown option
@@ -23,20 +32,37 @@ case $i in
 esac
 done
 
-if [ "$DANGER_REMOVE_CONTAINERS" == "true" ]; then
+if [ "$STOP_CONTAINERS" == "true" ]; then
   echo "###########################################"
-  echo "# [DANGER] remove domifa containers"
+  echo "# [WARN] STOP domifa containers"
   echo "###########################################"
-  # stop and remove all domifa containers
+  # stop all domifa containers
   (set -x && APP_DIR=$(pwd) docker-compose --project-name domifa --env-file ./.env -f ./docker-compose.local.yml down)
-  (set -x && APP_DIR=$(pwd) docker-compose --project-name domifa --env-file ./.env -f ./docker-compose.local.yml rm)
-fi
-if [ "$DANGER_DROP_VOLUMES" == "true" ]; then
-  echo "###########################################"
-  echo "# [DANGER] remove domifa volumes"
-  echo "###########################################"
-  # DANGER!!! purge all domifa volumes
-  (set -x && docker volume rm $(docker volume ls -f dangling=true -q | grep domifa))
+  if [ "$REMOVE_CONTAINERS" == "true" ]; then
+    echo "###########################################"
+    echo "# [WARN] REMOVE domifa containers"
+    echo "###########################################"
+    # remove all domifa containers
+    (set -x && APP_DIR=$(pwd) docker-compose --project-name domifa --env-file ./.env -f ./docker-compose.local.yml rm)
+    
+    if [ "$DANGER_DROP_VOLUMES" == "true" ]; then
+      echo "###########################################"
+      echo "# [DANGER] DROP domifa volumes"
+      echo "###########################################"
+      read -r -p "All 'domifa' volumes data WILL BE LOST. Are you sure? [y/N] " response
+      case "$response" in
+          [yY][eE][sS]|[yY])
+            # DANGER!!! purge all domifa volumes
+            (set -x && docker volume rm $(docker volume ls -f dangling=true -q | grep domifa))
+          ;;
+        *)
+              echo "# [SKIP]"
+              ;;
+      esac
+    fi
+    
+  fi
+  exit 0
 fi
 
 echo ""
@@ -44,8 +70,14 @@ echo "###########################################"
 echo "# START DOCKER ENV..."
 echo "#"
 echo ""
-# start mongo+postgres (with initial dumps)
+
+if [ "$WITH_DEV_CONTAINERS" == "true" ]; then
+# start backend + frontend + mongo + postgres (with initial dumps)
 (set -x && APP_DIR=$(pwd) docker-compose --project-name domifa --env-file ./.env -f ./docker-compose.local.yml up --build --detach --force-recreate)
+else
+# start mongo + postgres only (with initial dumps)
+(set -x && APP_DIR=$(pwd) docker-compose --project-name domifa --env-file ./.env -f ./docker-compose.local.yml up --build --detach --force-recreate mongo postgres)
+fi
 
 (set -x && docker ps -a)
 
@@ -53,20 +85,14 @@ echo ""
 echo "###########################################"
 echo "# TO RUN BACKEND:"
 echo ""
-echo "# enter container"
 echo "docker exec -it domifa-backend bash"
-echo ""
-echo "# start app"
-echo "yarn start"
+echo "yarn start # packages/backend"
 echo ""
 
 echo ""
 echo "###########################################"
 echo "# TO RUN FRONTEND:"
 echo ""
-echo "# enter container"
 echo "docker exec -it domifa-frontend bash"
-echo ""
-echo "# start app"
-echo "yarn start"
+echo "yarn start # packages/frontend"
 echo ""
